Fix refs. wrongly categorized as possibly omni
I thought "fat NFS" = animal fat, but as it turns out, if you follow it
all the way to the end, it always resolves to butter or vegetable
oils/margarine, so it's vegetarian. This only came up now because in the
previous commit I accidentally fixed a bug that caused input food
resolution to stop prematurely.